+++ layers/Engine/packages/loosecontracts/src/LooseContractManager.php
@@ -9,52 +9,11 @@ class LooseContractManager implements LooseContractManagerInterface
     /**
      * @var string[]
      */
-    protected array $requiredHooks = [];
-    /**
-     * @var string[]
-     */
-    protected array $implementedHooks = [];
-    /**
-     * @var string[]
-     */
     protected array $requiredNames = [];
     /**
      * @var string[]
      */
     protected array $implementedNames = [];
-
-    /**
-     * @return string[]
-     */
-    public function getNotImplementedRequiredHooks(): array
-    {
-        return array_diff(
-            $this->requiredHooks,
-            $this->implementedHooks
-        );
-    }
-
-    /**
-     * @param string[] $hooks
-     */
-    public function requireHooks(array $hooks): void
-    {
-        $this->requiredHooks = array_merge(
-            $this->requiredHooks,
-            $hooks
-        );
-    }
-
-    /**
-     * @param string[] $hooks
-     */
-    public function implementHooks(array $hooks): void
-    {
-        $this->implementedHooks = array_merge(
-            $this->implementedHooks,
-            $hooks
-        );
-    }
 
     /**
      * @return string[]
