+++ layers/Engine/packages/loosecontracts/src/LooseContractManagerInterface.php
@@ -9,18 +9,6 @@ interface LooseContractManagerInterface
     /**
      * @return string[]
      */
-    public function getNotImplementedRequiredHooks(): array;
-    /**
-     * @param string[] $hooks
-     */
-    public function requireHooks(array $hooks): void;
-    /**
-     * @param string[] $hooks
-     */
-    public function implementHooks(array $hooks): void;
-    /**
-     * @return string[]
-     */
     public function getNotImplementedRequiredNames(): array;
     /**
      * @param string[] $names
